requirements.txt
python-pptx
pillow
pytesseract
google-generativeai
python-dotenv